code refine
