+++ Projects/main.js
@@ -1,7 +1,10 @@
-import { setScene, controlCamera, setSceneElementsTemp} from "./scripts/sceneManager.js";
+import { setScene, controlCamera} from "./scripts/sceneManager.js";
+import { setSceneElementsTemp } from "./scripts/gridModels.js";
+import { addBlock } from "./scripts/buttonInteract.js";
 
 setScene();
 controlCamera();
+
 setSceneElementsTemp();
 
-// const uiManager = new UIManager(); // UIManager 인스턴스 생성
+addBlock();
